fix: apply the LNURL host check to page-advertised LNURLs and trim callback replies
isPrivateHost moves to common/lib/lnurl.ts so the publisher widget in the
popup, which loads the LNURL a page advertises in its meta tags, runs the
same check as the webln.lnurl entry point.

The channel and withdraw prompts and the lnurl-auth action now reply with
only the LNURL response fields (status, reason) instead of the raw callback
body.
