+++ src/app/screens/LNURLWithdraw/index.tsx
@@ -80,7 +80,7 @@ function LNURLWithdraw() {
         // ATTENTION: if this LNURL is called through `webln.lnurl` then we immediately return and return the response. This closes the window which means the user will NOT see the above successAction.
         // We assume this is OK when it is called through webln.
         if (navState.isPrompt) {
-          msg.reply(response.data);
+          msg.reply({ status: response.data.status });
         }
       } else {
         throw new Error(response.data.reason);
